pyproject.toml
[project]
name = "extremeweatherbench"
version = "1.0.1"
description = "Benchmarking weather and weather AI models using extreme events"
keywords = [
    "weather",
    "extreme events",
    "benchmarking",
    "forecasting",
    "climate",
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "dacite>=1.8.1",
    "gcsfs>=2024.12.0",
    "geopandas>=1.0.1",
    "joblib>=1.4.2",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pyyaml>=6.0.2",
    "regionmask>=0.13.0",
    "s3fs>=2024.12.0",
    "scores>=2.4.0",
    "tqdm>=4.67.1",
    "xarray[io]>=2025.7.1",
    "zarr>=3.0.9,<4",
    "kerchunk[dev]",
    "flox>=0.10.0",
    "click>=8.1.8",
    "polars>=1.32.1",
    "eccodes==2.42.0",
    "frozenlist==1.5.0",
    "pyogrio==0.10.0",
    "scikit-image>=0.25.2",
    "cartopy==0.24.1",
    "numba>=0.63.1",
    "sparse>=0.17.0",
    "icechunk>=1.1.14",
    "dask[complete]>=2025.1.0",
    "distributed>=2025.1.0",
]

[project.optional-dependencies]
data-prep = [
    "fsspec>=2024.12.0",
    "matplotlib>=3.10.0",
    "obstore>=0.8.2",
    "scipy>=1.13",
    "seaborn>=0.13.2",
    "ujson>=5.10.0",
    "virtualizarr==2.1.2",
    "dask[complete]>=2025.1.0",
    "distributed>=2025.1.0",
]

[dependency-groups]
dev = [
    "deptry>=0.23.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "isort>=6.0.1",
    "jupyterlab>=4.4.5",
    "lxml-stubs>=0.5.1",
    "mypy>=1.14.1",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.4",
    "types-pytz>=2025.2.0.20250809",
    "types-pyyaml>=6.0.12.20241230",
    "types-tqdm>=4.67.0.20250809",
    "python-semantic-release>=10.3.0",
    "twine>=5.1.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-autoapi[python]>=0.4.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=1.0.0",
    "pymdown-extensions>=10.19.1",
]

complete = ["extremeweatherbench[data-prep]"]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extremeweatherbench"]

[tool.hatch.build.targets.sdist]
include = ["src/extremeweatherbench/**/*"]

[project.urls]
Documentation = "https://extremeweatherbench.readthedocs.io/"
Repository = "https://github.com/brightbandtech/extremeweatherbench"


[project.scripts]
ewb = "extremeweatherbench.evaluate_cli:cli_runner"

[tool.uv.sources]
extremeweatherbench = { workspace = true }
kerchunk = { git = "https://github.com/fsspec/kerchunk" }

[tool.ruff]
line-length = 88
indent-width = 4
lint.ignore = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
case-sensitive = true

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
dist_path = "dist/"
upload_to_pypi = false
remote = { type = "github" }
commit_author = "semantic-release <semantic-release@python>"
commit_parser = "conventional"
commit_parser_options = { parse_squash_commits = "false", parse_merge_commits = "true" }
minor_tag = "[minor]"
patch_tag = "[patch]"
major_tag = "[major]"
build_command = """
  uv lock --offline
  git add uv.lock
  uv build
"""
# Only create GitHub releases for the current version, not historical ones
github_release_mode = "latest"
# Ensure assets are only uploaded for the current release, not past ones
upload_assets_for_all_releases = false

[tool.pytest]
addopts = ["--ignore=tests/test_golden.py", "--cov=extremeweatherbench"]
markers = [
    "integration: marks tests as integration tests (may be slow)",
    "slow: marks tests as slow (may take longer to complete)",
]
